AS Autobinder: remove unused ASClassBinder; Rename ASClassBinderNew to ASClassBinder
